import members: cleanup
