--- runtime-modules/membership/src/genesis_member.rs
@@ -1,12 +1,0 @@
-use serde::{Deserialize, Serialize};
-
-#[derive(Serialize, Deserialize)]
-pub struct Member<MemberId, AccountId, Moment> {
-    pub member_id: MemberId,
-    pub root_account: AccountId,
-    pub controller_account: AccountId,
-    pub handle: String,
-    pub avatar_uri: String,
-    pub about: String,
-    pub registered_at_time: Moment,
-}
